+++ gii/templates/modules/simple/Events.php
@@ -1,6 +1,6 @@
 <?php /* @var $generator \humhub\modules\devtools\gii\generators\ModuleGenerator */?><?= "<?php\n"; ?>
 
-namespace  <?= $generator->getClassNamespace(); ?>;
+namespace  <?= $generator->getClassNamespace() ?>;
 
 use Yii;
 use yii\helpers\Url;
@@ -15,11 +15,11 @@ class Events
     public static function onTopMenuInit($event)
     {
         $event->sender->addItem([
-            'label' => '<?= ucfirst($generator->moduleID); ?>',
+            'label' => '<?= ucfirst($generator->moduleID) ?>',
             'icon' => '<i class="fa <?= $generator->icon ?>"></i>',
-            'url' => Url::to(['/<?= $generator->moduleID; ?>/index']),
+            'url' => Url::to(['/<?= $generator->moduleID ?>/index']),
             'sortOrder' => 99999,
-            'isActive' => (Yii::$app->controller->module && Yii::$app->controller->module->id == '<?= $generator->moduleID; ?>' && Yii::$app->controller->id == 'index'),
+            'isActive' => (Yii::$app->controller->module && Yii::$app->controller->module->id == '<?= $generator->moduleID ?>' && Yii::$app->controller->id == 'index'),
         ]);
     }