+++ views/showcase/tabs/forms_simple/php.php
@@ -28,7 +28,7 @@ use humhub\modules\content\widgets\richtext\RichText;
 use yii\widgets\DetailView;
 use humhub\modules\user\widgets\Image as UserImage;
 
-/* @var $this \humhub\components\View */
+/* @var $this \humhub\modules\ui\view\components\View */
 /* @var $model TestModel */
 ?>