+++ views/showcase/tabs/richtext_custom_preset/try.php
@@ -12,7 +12,7 @@ use humhub\modules\content\widgets\richtext\RichTextField;
 use yii\bootstrap\ActiveForm;
 use yii\helpers\Url;
 
-/* @var $this \humhub\components\View */
+/* @var $this \humhub\modules\ui\view\components\View */
 ?>
 
 <script>