+++ widgets/views/panelRow.php
@@ -1,7 +1,13 @@
 <div class="row">
     <?php foreach ($items as $item) : ?>
         <div class="col-md-4 section<?= (isset($item['disabled']) && $item['disabled']) ? ' disabled' : ''?>">
-            <a href="<?= $item['url'] ?>"><?= $item['title'] ?></a>
+            <?php if(!isset($item['disabled']) || !$item['disabled']) : ?>
+            <a href="<?= $item['url'] ?>">
+                <?= $item['title'] ?>
+            </a>
+            <?php else : ?>
+                <?= $item['title'] ?>
+            <?php endif; ?>
             <p><?= $item['text'] ?></p>
         </div>
     <?php endforeach; ?>