+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 0.4.0
 ---------------------
 - Fix #20 Broken module generator layout
+- Enh: Define full module namespace instead of just prefix
 
 0.3.1
 ---------------------